+++ src/Game/Factories/Character/AIComponent/AIComponent.hpp
@@ -19,7 +19,7 @@ namespace Component
             AIComponent();
             virtual ~AIComponent() = default;
 
-            std::pair<size_t, size_t> getVelocity();
+            std::pair<double, double> getVelocity();
             bool putBomb();
             void setEnv(const std::shared_ptr<DataMatrix> &map, std::pair<size_t, size_t> pos, std::vector<std::pair<size_t, size_t>> &enemy);
